add: useUser
