+++ src/App.tsx
@@ -3,10 +3,12 @@ import { RecoilRoot } from 'recoil';
 import { ApolloClient, ApolloProvider } from '@apollo/client';
 import makeApolloClient from 'lib/apollo';
 import WithProvider from './WithProvider';
+import useIsFirstRender from 'hooks/useIsFirstRender';
 
 type CacheShape = any;
 
 function App() {
+  const isFirstRender = useIsFirstRender();
   const [client, setClient] = useState<ApolloClient<CacheShape> | null>(null);
 
   const fetchSession = async () => {
@@ -15,8 +17,10 @@ function App() {
   };
 
   useEffect(() => {
+    if (!isFirstRender) return;
+
     fetchSession();
-  });
+  }, [isFirstRender]);
 
   if (!client) {
     return null;
